+++ src/components/sidebar/TagMenu.js
@@ -2,18 +2,21 @@ import React from 'react'
 
 import TagMenuItem from './TagMenuItem'
 
-const TagMenu = ({ tags, selectedTag, ui }) => {
-  return (
-    <div className="tag-menu">
-      {tags.map(tag =>
-        <TagMenuItem
-          tag={tag}
-          key={tag.id}
-          active={tag.id === selectedTag}
-          ui={ui}
-        />
-       )}
-    </div>
-  )
+class TagMenu extends React.PureComponent {
+  render () {
+    const { tags, selectedTag } = this.props
+    return (
+      <div className="tag-menu">
+        {tags.map(tag =>
+          <TagMenuItem
+            tag={tag}
+            key={tag.id}
+            active={tag.id === selectedTag}
+          />
+         )}
+      </div>
+    )
+  }
 }
+
 export default TagMenu